Makes screw smaller, and removes lapping tool
The screw needed to be smaller to make it easier to screw in.

The lapping helper tool is not needed anymore because
we don't do lapping anymore.
We don't do the lapping because it was time consuming,
and did not improve precision noticably.

More detailed story:
We've started spray coating markers with retro-reflective
paint, and the layer of paint is easy to sand with small forces applied.
Using the lapping tool when sanding painted markers led to too
big forces being applied, and all the paint being scraped off.

That was before we discontinued lapping altogether.